// src/app/Type/type.tsx
export interface IWorkOuts {
  "id": number;
  "name": string;
  "image": string;
  "muscleGroups": string[];
  "equipment": string;
  "difficulty": string;
  "duration": number;
  "caloriesBurned": number;
  "sets": number;
  "reps": string;
  "rating": number;
  "description": string;
  "instructions": string[];
}